added thoughtdata and basic thought get route. thoughts now display when linked to a user

+++ seed/seeds.js
@@ -1,13 +1,20 @@
 const connection = require('../config/connection');
-const { User } = require('../models');
+const { User, Thought } = require('../models');
 const userdata = require('./userdata.json');
+const thoughtsdata = require('./thoughtsdata.json');
 
 connection.on('error', (err) => err);
 
 connection.once('open', async() => {
     try {
-        await connection.dropCollection('users');
+        const userCollection = await connection.db.listCollections({ name: 'users' });
+        const thoughtCollection = await connection.db.listCollections({ name: 'thoughts' });
+        if (userCollection && thoughtCollection) {
+            await connection.dropCollection('users');
+            await connection.dropCollection('thoughts');
+        };
         await User.insertMany(userdata);
+        await Thought.insertMany(thoughtsdata);
         console.log('seeded');
         process.exit(0);
     } catch (err) {
